Add support for IPUMS integer variables in Stata
When implied decimals are set, it means we know exactly how much precision
to show: '.nf' where n is the number of implied decimals.
When the number is a double float in our metadata, we go with the
underlying value and the '.0g' (general) formatting rules in Stata.
Otherwise the variable is a integer in our metadata, and we therefore
use '.0f' for formatting.

This functionality was released in gem 1.1.6 but was never committed to
the repository, and was therefore lost when version 1.1.7 was released.